+++ Sources/Sentry/SentryScope.m
@@ -2,7 +2,6 @@
 #import "SentryAttachment+Private.h"
 #import "SentryBreadcrumb.h"
 #import "SentryDefines.h"
-#import "SentryEnvelopeItemType.h"
 #import "SentryEvent+Private.h"
 #import "SentryInternalDefines.h"
 #import "SentryLevelMapper.h"
@@ -608,7 +607,7 @@ NS_ASSUME_NONNULL_BEGIN
 
         // Span could be nil as we do the first check outside the synchronize
         if (span != nil) {
-            if (![event.type isEqualToString:SentryEnvelopeItemTypeTransaction] &&
+            if (![event.type isEqualToString:SentryEnvelopeItemTypes.transaction] &&
                 [span isKindOfClass:[SentryTracer class]]) {
                 event.transaction = [[(SentryTracer *)span transactionContext] name];
             }