+++ Sources/SentryCppHelper/SentryCppHelper.swift
@@ -1,0 +1,1 @@
+// No content needed, we just need this empty file