+++ Tests/SentryTests/Helper/SentrySerializationNilTests.m
@@ -1,4 +1,3 @@
-#import "SentryEnvelope.h"
 #import "SentrySerialization.h"
 #import <XCTest/XCTest.h>